add date picker cells & move data entry table files to addParticipant folder
